requirements.txt
fastapi
uvicorn
aiohttp
pillow
sqlalchemy
pydantic
psycopg2